+++ examples/wallet-import-export/src/components/ImportWallet.tsx
@@ -71,6 +71,7 @@ export function ImportWallet(props: ImportWalletProps) {
       alert("Encrypted bundle is empty.");
       return;
     }
+
     const response = await axios.post("/api/importWallet", {
       userId: props.userId,
       walletName,
